+++ e2e/src/test/kotlin/Steps.kt
@@ -9,4 +9,18 @@ class Steps {
             ?.stubFor(post(urlEqualTo("/graphql"))
                 .andMatching(GraphqlBodyMatcher.withRequestJson(json)).willReturn(ok()))
     }
+
+    @Step("クエリ<query>を受け取って200を返すスタブを登録する")
+    fun setupGraphqlQueryStub(query: String) {
+        Datastore.server()
+            ?.stubFor(post(urlEqualTo("/graphql"))
+                .andMatching(GraphqlBodyMatcher.withRequestQueryAndVariables(query)).willReturn(ok()))
+    }
+
+    @Step("クエリ<query>と変数<variables>を受け取って200を返すスタブを登録する")
+    fun setupGraphqlQueryAndVariables(query: String, variables: String) {
+        Datastore.server()
+            ?.stubFor(post(urlEqualTo("/graphql"))
+                .andMatching(GraphqlBodyMatcher.withRequestQueryAndVariables(query, variables)).willReturn(ok()))
+    }
 }